+++ web/components/Admin/RoomWidget.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import React, { useContext, useEffect } from 'react';
+import { useContext, useEffect } from 'react';
 
 import Link from 'next/link';
