+++ web/components/Admin/ServerInfo.tsx
@@ -1,0 +1,50 @@
+'use client';
+
+import React from 'react';
+
+import { Container, Grid2 as Grid, Paper, Typography } from '@mui/material';
+
+import WidgetTitle from '@/components/common/WidgetTitle';
+
+import { queries } from '@/lib/queries/server';
+
+export default function ServerInfo() {
+  const { data: serverInfo } = queries.getInfo();
+
+  return (
+    <Container maxWidth="lg" className="py-8">
+      <Grid container spacing={3}>
+        <Grid size={{ xs: 12, md: 4, lg: 3 }}>
+          <Paper className="p-4">
+            <>
+              <WidgetTitle>Memory Usage</WidgetTitle>
+              <Typography component="p" variant="h4">
+                {serverInfo?.usedMemoryMB || 0}MB
+              </Typography>
+            </>
+          </Paper>
+        </Grid>
+        <Grid size={{ xs: 12, md: 4, lg: 3 }}>
+          <Paper className="p-4">
+            <React.Fragment>
+              <WidgetTitle>Server Uptime</WidgetTitle>
+              <Typography component="p" variant="h4">
+                {Math.round((serverInfo?.upTimeSecs || 0) / 60)} minutes
+              </Typography>
+            </React.Fragment>
+          </Paper>
+        </Grid>
+        <Grid size={{ xs: 12, md: 4, lg: 3 }}>
+          <Paper className="p-4">
+            <React.Fragment>
+              <WidgetTitle>Server Version</WidgetTitle>
+              <Typography component="p" variant="h4">
+                v{serverInfo?.serverVersion}
+              </Typography>
+            </React.Fragment>
+          </Paper>
+        </Grid>
+      </Grid>
+    </Container>
+  );
+}
