+++ web/context/ws.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { createContext } from 'react';
 
 import io from 'socket.io-client';
