--- web/lib/client.ts
@@ -1,8 +1,0 @@
-import axios from 'axios';
-
-const client = axios.create({
-  baseURL: process.env.NEXT_PUBLIC_API_URL || 'http://localhost:3000',
-  withCredentials: true,
-});
-
-export default client;
