+++ web/lib/queries/roles.ts
@@ -1,38 +1,7 @@
 import { keepPreviousData, useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 
-import client from '@/lib/client';
-
-export interface NewRole {
-  auth: string;
-  name: string;
-  role: string;
-}
-
-export interface PlayerRole {
-  auth: string;
-  name: string;
-  role: string;
-}
-
-export enum PlayerRoleEventType {
-  addRole = 'addrole',
-  removeRole = 'rmrole',
-  updateRole = 'updaterole',
-}
-
-export interface PlayerRoleEvent {
-  type: PlayerRoleEventType;
-  auth: string;
-  name: string;
-  role: string;
-  timestamp: number;
-}
-
-type Pagination = {
-  page: number;
-  pagingCount: number;
-  searchQuery?: string;
-};
+import { addRole, deleteRole, getPlayersRoleEvents, getPlayersRoles, updateRole } from '@/lib/api/roles';
+import { PlayerRole, PlayerRoleEvent, RolePagination } from '@/lib/types/roles';
 
 const queryKeys = {
   roles: ['roles'],
@@ -40,38 +9,17 @@ const queryKeys = {
 };
 
 const queries = {
-  getPlayersRoles: ({ page, pagingCount, searchQuery = '' }: Pagination) =>
+  getPlayersRoles: (pagination: RolePagination) =>
     useQuery<PlayerRole[]>({
-      queryKey: [...queryKeys.roles, { page, pagingCount, searchQuery }],
-      queryFn: async () => {
-        const index: number = (page - 1) * pagingCount;
-        try {
-          const result = await client.get(
-            `/api/v1/roleslist?searchQuery=${searchQuery}&start=${index}&count=${pagingCount}`,
-          );
-          return result.data;
-        } catch (e: any) {
-          throw new Error('Failed to load roles list.');
-        }
-      },
+      queryKey: [...queryKeys.roles, pagination],
+      queryFn: () => getPlayersRoles(pagination),
       placeholderData: keepPreviousData,
     }),
 
-  getPlayersRoleEvents: ({ page, pagingCount, searchQuery = '' }: Pagination) =>
+  getPlayersRoleEvents: (pagination: RolePagination) =>
     useQuery<PlayerRoleEvent[]>({
-      queryKey: [...queryKeys.rolesEvents, { page, pagingCount, searchQuery }],
-      queryFn: async () => {
-        const index: number = (page - 1) * pagingCount;
-        try {
-          const result = await client.get(
-            `/api/v1/roleslist/events?searchQuery=${searchQuery}&start=${index}&count=${pagingCount}`,
-          );
-          console.log(`events: ${result.data}`);
-          return result.data;
-        } catch (e: any) {
-          throw new Error('Failed to load roles events list.');
-        }
-      },
+      queryKey: [...queryKeys.rolesEvents, pagination],
+      queryFn: () => getPlayersRoleEvents(pagination),
       placeholderData: keepPreviousData,
     }),
 };
@@ -81,21 +29,7 @@ const mutations = {
     const queryClient = useQueryClient();
 
     return useMutation({
-      mutationFn: async (newRole: NewRole) => {
-        try {
-          await client.post(`/api/v1/roleslist/${newRole.auth}`, {
-            name: newRole.name,
-            role: newRole.role,
-          });
-        } catch (error: any) {
-          if (error.response.status === 409) {
-            throw new Error(`Player '${newRole.name}' (id: ${newRole.auth}) already added.`);
-          } else {
-            throw new Error(`Failed to add ${newRole.name} (id: ${newRole.auth}).`);
-          }
-        }
-      },
-
+      mutationFn: addRole,
       onSettled: () => {
         queryClient.invalidateQueries({ queryKey: queryKeys.roles });
         queryClient.invalidateQueries({ queryKey: queryKeys.rolesEvents });
@@ -107,17 +41,7 @@ const mutations = {
     const queryClient = useQueryClient();
 
     return useMutation({
-      mutationFn: async (role: PlayerRole) => {
-        try {
-          await client.put(`/api/v1/roleslist/${role.auth}`, {
-            name: role.name,
-            role: role.role,
-          });
-        } catch (error: any) {
-          throw new Error(`Failed to update ${role.name}'s role.`);
-        }
-      },
-
+      mutationFn: updateRole,
       onSettled: () => {
         queryClient.invalidateQueries({ queryKey: queryKeys.roles });
         queryClient.invalidateQueries({ queryKey: queryKeys.rolesEvents });
@@ -129,14 +53,7 @@ const mutations = {
     const queryClient = useQueryClient();
 
     return useMutation({
-      mutationFn: async ({ auth, name }: { auth: string; name: string }) => {
-        try {
-          await client.delete(`/api/v1/roleslist/${auth}?name=${name}`);
-        } catch (error: any) {
-          throw new Error(`Failed to delete ${name}'s role.`);
-        }
-      },
-
+      mutationFn: deleteRole,
       onSettled: () => {
         queryClient.invalidateQueries({ queryKey: queryKeys.roles });
         queryClient.invalidateQueries({ queryKey: queryKeys.rolesEvents });
@@ -145,4 +62,4 @@ const mutations = {
   },
 };
 
-export { queryKeys, queries, mutations };
+export { mutations, queries, queryKeys };
