+++ web/lib/queries/server.ts
@@ -1,5 +1,6 @@
-import client from '../client';
 import { useQuery } from '@tanstack/react-query';
+
+import { getServerInfo } from '@/lib/api/server';
 
 const queryKeys = {
   info: ['server', 'info'],
@@ -9,15 +10,8 @@ const queries = {
   getInfo: () =>
     useQuery({
       queryKey: queryKeys.info,
-      queryFn: async () => {
-        try {
-          const result = await client.get('/api/v1/system');
-          return result.data;
-        } catch (e: any) {
-          throw new Error('Failed to load server info.');
-        }
-      },
+      queryFn: getServerInfo,
     }),
 };
 
-export { queryKeys, queries };
+export { queries, queryKeys };
