+++ web/lib/types/server.ts
@@ -1,0 +1,5 @@
+export type ServerInfo = {
+  usedMemoryMB: number;
+  upTimeSecs: number;
+  serverVersion: string;
+};
